requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
customtkinter==4.5.10
darkdetect==0.7.1
decorator==5.1.1
Deprecated==1.2.13
firebase-admin==5.3.0
gcloud==0.18.3
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
MarkupSafe==2.1.1
msgpack==1.0.4
Naked==0.1.31
oauth2client==4.1.3
pika==1.2.1
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.21.0
python-jwt==3.3.4
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
retry==0.9.2
rsa==4.9
shellescape==3.8.1
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
